Added footer link to GitHub and darkmode switch. Also fixed responsive design issues and Download page copy
