+++ styles/components/Button.module.scss
@@ -34,3 +34,8 @@
   font-weight: bold;
   font-size: 1.5rem;
 }
+
+.icon {
+  padding: 0;
+  border-radius: 50%;
+}
