+++ styles/pages/Download.module.scss
@@ -1,0 +1,9 @@
+.copy {
+  width: 50%;
+}
+
+@media only screen and (max-width: 800px) {
+  .copy {
+    width: 90%;
+  }
+}
